container-service/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-socketio==5.4.0
python-engineio==4.3.0
docker==5.0.3
pyjwt==2.1.0
python-dotenv==0.19.0